// src/enum/common.enum.ts
export enum GenderEnum {
    MALE = 'Male',
    FEMALE = 'Female',
    OTHER = 'Other'
}
export enum SiteContentEnum {
    PRIVACYPOLICY = 'Privacy Policy',
    TERMCONDITION = 'Terms And Condition',
    ABOUTUS = 'About Us'
}
export enum SortOrderEnum {
    ASCENDING = 'asc',
    DESCENDING = 'desc'
}
export enum PlatformEnum {
    IOS = 'ios',
    ANDROID = 'android'
}
export enum CommunicationCategoryEnum {
    MARKETNEWS = 'Market News',
    PROFILESUGGESTION = 'Profile Suggestion'
}
export enum CommunicationTypeEnum {
    EMAIL = 'Email',
    PUSHNOTIFICATION = 'Push Notification'
}