// src/enum/event.enum.ts
export enum InquiryEventEnum {
    CREATED = 'inquiry.created',
    UPDATED = 'inquiry.updated',
    CLOSED = 'inquiry.closed'
}
export enum PaymentEventEnum {
    ACCEPTED = 'payment.accepted',
    REJECTED = 'payment.rejected'
}
export enum SignalEventEnum {
    CREATED = 'signal.created',
    UPDATED = 'signal.updated',
    COMMENT = 'signal.coment'
}
export enum ProfileEventEnum {
    STATUS = 'status.updated'
}
export enum TransactionEventEnum {
    STATUS = 'status.updated'
}